Fix PA4 on dept machines

+++ project/src/depth_image_list.cpp
@@ -16,11 +16,11 @@ DepthImageList DepthImageList::fromFile(std::string filename) {
 
 const std::vector<DepthImageList::Item>& 
 DepthImageList::calculateNearestNeighbors() {
-    for(int i = 0; i < list.size(); ++i) {
+    for(unsigned int i = 0; i < list.size(); ++i) {
         auto& current = list.at(i);
-        for(int j = i + 1; j < list.size(); ++j) {
+        for(unsigned int j = i + 1; j < list.size(); ++j) {
             auto& test = list.at(j);
-            if(test.nearestNeighborIndex == i) {
+            if(test.nearestNeighborIndex == static_cast<int>(i)) {
                 continue;
             }
             double sum = current.img.minimumSumComparison(test.img);
